Allow for installable themes,
This change adds a method `get_theming_root` to `Theme` and calls that passing
the theming root if necessary. It returns either a list of full paths to
themes within virtual environments (allowing them to be used) or a single
full path if `theme_slug` is passed in.

Update docs to explain usage.